pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Coleta via API do OpenProject e exportacao para Power BI"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv",
  "requests",
  "pandas",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
